fix: harden the vendor/AIDA trust boundary
Security-audit findings:

HIGH: a nil sensor-list element (<SensorItem i:nil="true"/>) deserialized to
null and the projection threw an NRE that reached the render tick uncaught
(compose is try/finally only and the pump has no catch), killing the process
every launch and every inspector open. The projection now skips nulls.

Defense in depth: SkiaFrameCompositor wraps each widget's Render in a catch
that logs once per placement and keeps compositing, so no single widget can
take down the frame.

MEDIUM: MemoryMappedAidaMmapSource compared offset + length in 32-bit, so a
producer-controlled offset near int.MaxValue wrapped negative and bypassed the
range guard, forcing a close/reopen churn while reporting a read failure. The
compare is now 64-bit.

LOW: the inspector's sensor labels (untrusted Name/Type, bounded only by the
8 KB per-string quota) are truncated to 128 chars before they become WPF
combo-item text.

+++ ModernWigiDash.Hardware/Aida64/MemoryMappedAidaMmapSource.cs
@@ -103,7 +103,10 @@ public sealed class MemoryMappedAidaMmapSource : IAidaMmapSource
                 return false;
             }
 
-            if (offset + length > _accessor!.Capacity)
+            // 64-bit compare: a producer-controlled offset near int.MaxValue made
+            // the 32-bit sum wrap negative and pass this guard, leaving the
+            // framework to reject the read (a per-frame close/reopen churn).
+            if ((long)offset + length > _accessor!.Capacity)
             {
                 error = "AIDA64 map too small for the requested range";
                 return false;
